starters/computer-vision/requirements.txt
# web framework
flask

# System variables
python-dotenv

# Azure SDK components
azure-cognitiveservices-vision-computervision
